requirements.txt
absl-py==0.9.0
astor==0.8.0
attrs==19.3.0
backcall==0.2.0
bleach==3.1.5
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1594357566944/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
ipykernel @ file:///C:/ci/ipykernel_1594829155762/work/dist/ipykernel-5.3.2-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593433781269/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
Markdown==3.1.1
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
notebook==6.0.3
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
packaging==20.4
pandas @ file:///C:/ci/pandas_1592841747005/work
pandocfilters==1.4.2
parso==0.7.0
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow @ file:///C:/ci/pillow_1594298234712/work
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.16.0
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.0
scipy @ file:///C:/ci/scipy_1592916958183/work
Send2Trash==1.5.0
six==1.15.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-hub==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
